+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Vite
 
 - disable code splitting for standalone bundles
+- make build plugin config take precedence over user config
 
 ## 0.1.0 (2026-06-10)